experementing with touch and hover
