+++ touch-input_with_video.py
@@ -12,22 +12,27 @@ ToDo:
 '''
 
 import cv2
-import sys
 from os import path
-import uuid
 import time
-import numpy as np
 
 from collections import deque
 
+# own helper classes
+from helper_classes.agglomerative_clustering import Agglomerative_Cluster_Model
+
 CURR_DIR = path.dirname(__file__)
 #RECORDINGS_PATH = path.join(CURR_DIR, f"assets/{str(uuid.uuid4())}.mp4")
-RECORDED_SESSION = path.join(CURR_DIR, f"assets/touch_stärker_2.mp4")
+#RECORDED_SESSION = path.join(CURR_DIR, f"assets/touch_stärker_2.mp4") # Touch
+RECORDED_SESSION = path.join(CURR_DIR, f"assets/touch_hover_leichter_2.mp4") # Hover
+
+# area treshholds for touch/hover detection
+area_lower_limit = 0
+area_upper_limit = 50
 
 # Create a video capture object for the webcam
 cap = cv2.VideoCapture(RECORDED_SESSION)
 
-cutoff = 25
+cutoff = 15 # 15 für touch
 radius_touch = 40
 radus_hover = 60
 
@@ -36,50 +41,7 @@ dom_fing_dq:list = deque(maxlen=1)
 # current dominant fingertip touch
 curr_dom_touch:list = deque(maxlen=1)
 
-# Agglomerative Clustering
-# https://cullensun.medium.com/agglomerative-clustering-for-opencv-contours-cd74719b678e
-def calculate_contour_distance(contour1, contour2): 
-    x1, y1, w1, h1 = cv2.boundingRect(contour1)
-    c_x1 = x1 + w1/2
-    c_y1 = y1 + h1/2
-
-    x2, y2, w2, h2 = cv2.boundingRect(contour2)
-    c_x2 = x2 + w2/2
-    c_y2 = y2 + h2/2
-
-    return max(abs(c_x1 - c_x2) - (w1 + w2)/2, abs(c_y1 - c_y2) - (h1 + h2)/2)
-
-# Agglomerative Clustering
-# https://cullensun.medium.com/agglomerative-clustering-for-opencv-contours-cd74719b678e
-def merge_contours(contour1, contour2):
-    return np.concatenate((contour1, contour2), axis=0)
-
-# Agglomerative Clustering
-# https://cullensun.medium.com/agglomerative-clustering-for-opencv-contours-cd74719b678e
-def agglomerative_cluster(contours, threshold_distance=40.0):
-    current_contours = contours
-    while len(current_contours) > 1:
-        min_distance = None
-        min_coordinate = None
-
-        for x in range(len(current_contours)-1):
-            for y in range(x+1, len(current_contours)):
-                distance = calculate_contour_distance(current_contours[x], current_contours[y])
-                if min_distance is None:
-                    min_distance = distance
-                    min_coordinate = (x, y)
-                elif distance < min_distance:
-                    min_distance = distance
-                    min_coordinate = (x, y)
-
-        if min_distance < threshold_distance:
-            index1, index2 = min_coordinate
-            current_contours[index1] = merge_contours(current_contours[index1], current_contours[index2])
-            del current_contours[index2]
-        else: 
-            break
-
-    return current_contours
+aggl_clusterer = Agglomerative_Cluster_Model()
 
 
 def touch_areas(img_raw):
@@ -100,14 +62,14 @@ def touch_areas(img_raw):
 
     for contour in contours:
         area = cv2.contourArea(contour)
-        if area >= 10 and area <= 50:
+        if area >= area_lower_limit and area <= area_upper_limit:
             #(x,y),radius = cv2.minEnclosingCircle(contour)
             #center = (int(x),int(y))
             #radius = int(radius) * 3
-            #cv2.circle(img_gray,center,radius,(0,255,0),3)
+            #cv2.circle(img_gray,center,radius,(0,255,0),3)q
             touch_areas_contours.append(contour)
             
-    area_contours_clustered = agglomerative_cluster(touch_areas_contours)
+    area_contours_clustered = aggl_clusterer.agglomerative_clustering(touch_areas_contours)
         
     final_areas:list = []
